change the attach target of bpf code
