+++ examples/uprobe/uprobe.bpf.c
@@ -5,14 +5,14 @@
 
 char LICENSE[] SEC("license") = "GPL";
 
-SEC("uprobe//proc/self/exe:uprobe_add")
+SEC("uprobe//proc/pid/exe:uprobe_add")
 int BPF_KPROBE(uprobe_add, int a, int b)
 {
 	bpf_printk("uprobed_add ENTRY: a = %d, b = %d", a, b);
 	return 0;
 }
 
-SEC("uprobe//proc/self/exe:uprobe_sub")
+SEC("uprobe//proc/pid/exe:uprobe_sub")
 int BPF_KPROBE(uprobe_sub, int a, int b)
 {
 	bpf_printk("uprobed_sub ENTRY: a = %d, b = %d", a, b);
