+++ examples/uprobe/uprobe.c
@@ -4,14 +4,6 @@
 #include "libbpf-wasm.h"
 #include "uprobe.skel.h"
 
-int uprobe_add(int a, int b){
-    return a + b;
-}
-
-int uprobe_sub(int a, int b){
-    return a - b;
-}
-
 int main(int argc, char* argv[]) {
     struct uprobe_bpf* skel = NULL;
     int err;
@@ -27,11 +19,9 @@ int main(int argc, char* argv[]) {
         printf("Failed to attach BPF skeleton\n");
         return -1;
     }
+    
     printf("Load and attach BPF uprobe successfully\n");
-    while (1) {
-        int a = 1;
-        int b = 2;
-        printf("%d %d\n", uprobe_add(a, b), uprobe_sub(a, b));
-        sleep(2);
+    while(1){
+        sleep(10);
     }
 }
